+++ lib/omniauth-quickbooks-oauth2-modern.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 require "omniauth/quickbooks_oauth2_modern/version"
+require "omniauth/quickbooks_oauth2_modern/token_client"
 require "omniauth/strategies/quickbooks_oauth2_modern"